feat: Move renderable mixin to main/utils.py
